style: rename data.py -> phases.py, per code review
