+++ tests/parser/functions/test_abi.py
@@ -3,11 +3,11 @@ import pytest
 from vyper.compiler import (
     compile_code,
 )
-from vyper.compiler.data import (
-    CompilerData,
-)
 from vyper.compiler.output import (
     build_abi_output,
+)
+from vyper.compiler.phases import (
+    CompilerData,
 )
 
 source_codes = [
